level 1 complete
boss spawns and slowly moves in. Should change wave time on turn in
though. current debug is like 1-10 seconds. Noticed the boss level didnt
have much going on, not sure if on purpose? will work on it a bit
tomorrow and see if I cant get a little bit of the boss fight working.